SimpleXML: better return annotation for xpath
If `xpath` is called from `SimpleXMLIterator`, which extends `SimpleXMLElement`, the result type of `xpath` would be `SimpleXMLIterator[]`, not `SimpleXMLElement[]`. Currently, Phpstorm would report non-existing `current`/`next`/rewind` methods that are called on `xpath` results even if the underlying type is `SimpleXMLIterator` (which has those methods). A simplest fix is to annotate return type as `static[]`, because this way Phpstorm is able to infer proper type of the elements and stops making false reports about undefined methods.